Switch opacity and scale to float to keep consistency
All ports which deals with scale are using as floats, WK API
also expects the values as floats. Opacity is also expected
as float, so it should not be a problem to switch it too.

Reviewed by: Caio Marcelo

+++ Source/WebKit2/UIProcess/API/nix/WebView.h
@@ -56,13 +56,13 @@ public:
     virtual void setDrawBackground(bool) = 0;
     virtual bool drawBackground() const = 0;
 
-    virtual void setScale(double) = 0;
-    virtual double scale() const  = 0;
+    virtual void setScale(float) = 0;
+    virtual float scale() const = 0;
 
     virtual WKSize visibleContentsSize() const = 0;
 
-    virtual void setOpacity(double) = 0;
-    virtual double opacity() const = 0;
+    virtual void setOpacity(float) = 0;
+    virtual float opacity() const = 0;
 
     virtual void paintToCurrentGLContext() = 0;
